requirements.txt
coverage==7.2.2
Django==4.1.7
django-admin-thumbnails==0.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
djoser==2.2.0
Pillow==9.4.0
django_debug_toolbar==3.8.1
djlint==1.19.16
flake8==6.0.0
gunicorn==20.1.0
psycopg2==2.9.5
uWSGI==2.0.21
